+++ src/types/db.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct Mod {
+    id: String,
+    name: String,
+    download_url: String,
+}
